Fugu clean-up: Fix mismatch warning in printf call
Depending on architecture, the return type of atomic64_read()
is changed.

Reviewed-on: https://android.intel.com/215614
Tested-by: Fourdan, Olivier <olivier.fourdan@intel.com>

+++ drivers/external_drivers/intel_media/display/tng/drv/mdfld_dsi_pkg_sender.c
@@ -1371,10 +1371,10 @@ int mdfld_dsi_send_dcs(struct mdfld_dsi_pkg_sender *sender,
 			atomic64_read(&sender->te_seq)) {
 			mutex_unlock(&sender->lock);
 			if (dev_priv->b_async_flip_enable)
-				DRM_INFO("reject WMS LSU[%ld], te_seq[%ld]\n",
-					 atomic64_read(&sender->
+				DRM_INFO("reject WMS LSU[%lld], te_seq[%lld]\n",
+					 (long long) atomic64_read(&sender->
 						       last_screen_update),
-					 atomic64_read(&sender->te_seq));
+					 (long long) atomic64_read(&sender->te_seq));
 			return -EAGAIN;
 		}
